Initial fetch refactor
